Consolidate strategy tests into reckoner
Prepare for refactoring to remove the strategy classes and build them
into the Reckoner. This puts all of the strategy tests at the Reckoner
level, so it ensures that they'll apply when the logic from the
strategies themselves is moved in there.
